requirements.txt
faiss-cpu==1.11.0
sentence-transformers==5.0.0
requests==2.32.4
pdfplumber==0.11.7
numpy==2.3.1
hf_xet==1.1.5
python-dotenv==1.1.1
transformers==4.53.3
